delete fp detect id range

+++ src/process_event.c
@@ -2629,16 +2629,24 @@ void process_event(void)
 							int8_t id;
 							
 							id = Get_Finger_User_From_InterIndex(finger_num);
-	
-							Delect_One_ID((uint8_t) id);
-							PASSWD_COMPARE_OK();
-							Delete_ID_Flash_Once(id);
-							if(Get_User_id_Number()==0)
+							if((id>0)&&(id<=USER_ID_MAX))
 							{
-								Lock_NULL_Indication();
+								Delect_One_ID((uint8_t) id);
+								PASSWD_COMPARE_OK();
+								Delete_ID_Flash_Once(id);
+								if(Get_User_id_Number()==0)
+								{
+									Lock_NULL_Indication();
+								}
+								Wait_Select_Delete_Mode(DELETE_USER_BY_FP);
+								Exit_Finger_Current_Operate();
 							}
-							Wait_Select_Delete_Mode(DELETE_USER_BY_FP);
-							Exit_Finger_Current_Operate();	
+							else
+							{
+									PASSWD_COMPARE_ERR();
+									Hal_SEG_LED_Display_Set(HAL_LED_MODE_ON, Lock_Enter_DELETE_USER_BY_FP() ); 
+							}
+	
 						}
 						else 
 						{	
@@ -2775,15 +2783,26 @@ void process_event(void)
 							int8_t id;
 							
 							id = Get_Finger_Admin_From_InterIndex(finger_num);
-	
-							Delect_One_ID((uint8_t) id);
-							PASSWD_COMPARE_OK();
-							Delete_ID_Flash_Once(id);
-							if(Get_Admin_id_Number()==0)
+							if((id>USER_ID_MAX)&&(id<=ADMIN_ID_MAX))
 							{
-								Set_Work_Mode(NORMAL);
-								Lock_NULL_Indication();
+								Delect_One_ID((uint8_t) id);
+								PASSWD_COMPARE_OK();
+								Delete_ID_Flash_Once(id);
+								if(Get_Admin_id_Number()==0)
+								{
+									Set_Work_Mode(NORMAL);
+									Lock_NULL_Indication();
+								}
+								Wait_Select_Delete_Mode(DELETE_ADMIN_BY_FP);
+								Exit_Finger_Current_Operate();
 							}
+							else
+							{
+								PASSWD_COMPARE_ERR();
+	//							delay_ms(300);
+								Hal_SEG_LED_Display_Set(HAL_LED_MODE_ON, Lock_Enter_DELETE_ADMIN_BY_FP() ); 
+							}
+								
 						}
 						else 
 						{	
@@ -2791,8 +2810,7 @@ void process_event(void)
 //							delay_ms(300);
 							Hal_SEG_LED_Display_Set(HAL_LED_MODE_ON, Lock_Enter_DELETE_ADMIN_BY_FP() ); 
 						}
-						Wait_Select_Delete_Mode(DELETE_ADMIN_BY_FP);
-						Exit_Finger_Current_Operate();
+
 					}
 					
 				}
